chore: extract secret-specific helpers
To keep those in `secrets` module.
